requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI Services
openai==1.3.7
tiktoken==0.11.0
numpy==2.3.3

# Caching
redis==5.0.1

# Configuration
python-decouple==3.8

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# File Storage
django-storages==1.14.2
boto3==1.29.7
Pillow==10.4.0

# File Processing
PyPDF2==3.0.1

# Development
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
